requirements.txt
Flask==1.1.2
flask-cors==3.0.10
python-dotenv==0.14.0
requests==2.25.1
